+++ src/core/shading.ts
@@ -1,14 +1,17 @@
 
-import { generateUUID } from "../math";
+import { generateUUID, Vector3, Matrix4 } from "../math";
 import { Nullable } from "../type";
 import { GLProgramConfig, GLProgram } from "../webgl/program";
 import { ShaderGraph } from "../shader-graph/shader-graph";
 import { Observable } from "./observable";
 import { RenderEngine } from "../engine/render-engine";
-import { UniformProxy } from "../engine/uniform-proxy";
+import { ShaderCommonUniformInputNode } from '../shader-graph/shader-node';
+import { GLDataType } from '../webgl/shader-util';
+import { Vector2 } from '../math/vector2';
+import { Vector4 } from '../math/vector4';
+import { uniform } from '../shader-graph/node-maker';
 
 export interface ShaderUniformProvider{
-  providerName: string;
 
   /**
   * impl this to decorate your shader source, add uniform input
@@ -65,4 +68,57 @@ export class Shading {
     engine.deleteProgram(this);
   }
 
+}
+
+
+
+export function MapUniform<T>(remapName: string) {
+  return (target: BaseEffectShading<T>, key: string) => {
+    let val: T = target[key];
+    const getter = () => {
+      return val;
+    };
+    const setter = (value: T) => {
+      target.uniforms.set(remapName, value);
+      val = value;
+    };
+
+    target.propertyUniformNameMap.set(key, remapName);
+
+    Object.defineProperty(target, key, {
+      get: getter,
+      set: setter,
+      enumerable: true,
+      configurable: true,
+    });
+  };
+}
+
+export abstract class BaseEffectShading<T> implements ShaderUniformProvider{
+
+  abstract decorate(graph: ShaderGraph): void;
+
+  propertyUniformNameMap: Map<string, string> = new Map();
+  uniforms: Map<string, any> = new Map();
+
+  getPropertyUniform(name: keyof T): ShaderCommonUniformInputNode {
+    const uniformName = this.propertyUniformNameMap.get(name as string);
+    const value = this[name as string];
+    if (value === undefined) {
+      throw "uniform value not given"
+    }
+    if (typeof value === "number") {
+      return uniform(uniformName, GLDataType.float).default(value);
+    } else if (value instanceof Vector2) {
+      return uniform(uniformName, GLDataType.floatVec2).default(value);
+    } else if (value instanceof Vector3) {
+      return uniform(uniformName, GLDataType.floatVec3).default(value);
+    } else if (value instanceof Vector4) {
+      return uniform(uniformName, GLDataType.floatVec4).default(value);
+    } else if (value instanceof Matrix4) {
+      return uniform(uniformName, GLDataType.Mat4).default(value);
+    } else {
+      throw "un support uniform value"
+    }
+  }
 }
